BackendWoekSense/requirements.txt
Django>=5.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
Pillow>=10.1.0
twilio>=8.10.0
geopy>=2.3.0
python-dateutil>=2.8.2
pytz>=2024.1
torch>=2.1.2
torchvision>=0.16.2
transformers>=4.35.2
datasets>=2.14.6
numpy>=1.24.3
scikit-learn>=1.3.2
matplotlib>=3.8.2
tqdm>=4.66.1
requests>=2.31.0
python-multipart>=0.0.6
